refactor(content): boot by injecting ac-root element
