+++ src/core/index.ts
@@ -1,3 +1,4 @@
+export * from './app';
 export * from './project';
 export * from './paper-project';
 export * from './inspector';
